+++ tests/testthat/test-names.R
@@ -11,3 +11,19 @@ test_that("names_switch() works", {
   )
 })
 
+test_that("names_sort() works", {
+  x <- set_names0(rep(NA, 3), c(-1, 10, 2))
+
+  expect_equal(
+    sort_names(x),
+    set_names0(rep(NA, 3), c(-1, 10, 2))
+  )
+
+  expect_equal(
+    sort_names(x, numeric = TRUE),
+    set_names0(rep(NA, 3), c(-1, 2, 10))
+  )
+
+  expect_error(sort_names(list(a = 1)))
+  expect_error(sort_names(NA))
+})