+++ NEWS.md
@@ -20,6 +20,7 @@
 * adds `tableNA()` to make a table from `NA` values
 * `round_by()` gains an additional argument `include0` which if `FALSE` will replace `0` values with `by`
 * `assign_labels.data.frame()` gains an additional argument `.missing` to set how to control for missing labels: you can now use a `warning` for a missing label (instead of an error) or silently ignore any missing labels
+* `sort_names()` gains a new argument `numeric` to try to sort names of `x` by their numeric value
 
 # mark 0.1.4